[Modified] useFetch에 Generic Type을 추가함
useFetch가 다루는 데이터에 대한 의존성을 없애기 위해 Generic Type 사용하도록 변경
