+++ client/src/components/FeedWorks/index.tsx
@@ -1,17 +1,8 @@
-import React, { useState, useEffect } from 'react';
-import styled from 'styled-components';
-import axios from 'axios';
+import React, { useEffect } from 'react';
 import Card from '../Card';
 import useFetch from '../../hooks/useFetch';
 import { API_SERVER } from '../../utils/constants';
-
-const Container = styled.div`
-    display: flex;
-    flex-flow: wrap;
-    justify-content: space-between;
-    width: 60%;
-    margin: auto;
-`;
+import * as S from './styles';
 
 interface IImage{
   _id: string;
@@ -23,50 +14,34 @@ interface IImage{
   url: string;
 }
 const FeedWorks:React.FC = () => {
-  const [data, setData] = useState<IImage[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [isError, setIsError] = useState(false);
+  const [{
+    data, isLoading, isError,
+  }, doFetch] = useFetch<IImage>([]);
 
   useEffect(() => {
-    async function fetchData() {
-      setIsLoading(true);
-      setIsError(false);
-      try {
-        const result = await axios(`${API_SERVER}/image/`);
-        if (!result.data.success) {
-          alert(result.data.data.message);
-          setIsError(true);
-          setIsLoading(false);
-        } else {
-          const images = result.data.data;
-          const newData = data.concat(images);
-          setData(newData);
-          setIsLoading(false);
-        }
-      } catch (e) {
-        setIsError(true);
-      }
-    }
-    fetchData();
+    doFetch(`${API_SERVER}/image/`);
   }, []);
 
   return (
-    <Container>
-      {isError && !isLoading && <div>Something wrong...</div>}
-      {isLoading ? (
-        <div>Loading...</div>
-      ) : (
-        data.map((image) => (
-          <Card
-            imageId={image._id}
-            imgUrl={image.url}
-            creator={image.creator}
-            key={image._id}
-          />
-        ))
-      )}
+    <S.Container>
+      {
+        isError
+          ? (<div>Something wrong...</div>)
+          : isLoading
+            ? (<div>Loading...</div>)
+            : (
+              data.map((image) => (
+                <Card
+                  imageId={image._id}
+                  imgUrl={image.url}
+                  creator={image.creator}
+                  key={image._id}
+                />
+              ))
+            )
+      }
 
-    </Container>
+    </S.Container>
   );
 };
