+++ client/src/components/FeedWorks/styles.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+    display: flex;
+    flex-flow: wrap;
+    justify-content: space-between;
+    width: 60%;
+    margin: auto;
+`;
